refactor(eth): move all ENS code into ens.go
Gather the ENS declarations scattered across contract.go, dataparser.go,
ethparser.go, and ethrpc.go into a single ens.go: the NameRegistered signature
and ENS contract addresses/selectors, getEnsRecord, SetEnsSuffix,
UseEnsReverseAliases, and the forward path (ensNameHash, keccak256,
parseENSAddressFromResult, ensContracts, ResolveENS, CheckENSExpiration).

Pure move, no behavior change; struct fields (EnsSuffix, DisableEnsReverse,
DisableEnsAliases) stay with their types.
