+++ bchain/coins/eth/dataparser.go
@@ -303,37 +303,4 @@ func (p *EthereumParser) ParseInputData(signatures *[]bchain.FourByteSignature, 
 	return &parsed
 }
 
-// WARNING: not production-ready, disabled by default (see UseEnsReverseAliases).
-// Trusts any log emitter (spoofable labels), does not validate names, and ignores
-// expiry/ownership (stale/duplicate labels). Use forward ResolveENS instead.
-//
-// getEnsRecord parses an ENS record from a transaction log entry.
-func getEnsRecord(l *rpcLogWithTxHash) *bchain.AddressAliasRecord {
-	if len(l.Topics) == 3 && l.Topics[0] == nameRegisteredEventSignature && len(l.Data) >= 322 {
-		address, err := addressFromPaddedHex(l.Topics[2])
-		if err != nil {
-			return nil
-		}
-		c, err := strconv.ParseInt(l.Data[194:194+64], 16, 64)
-		if err != nil {
-			return nil
-		}
-		const nameStart = 194 + 64
-		// int(c)<<1 can overflow: c is attacker-controlled (up to 2^63-1) and Go's
-		// signed left shift wraps silently, so de can land anywhere in [0, 257]
-		// below nameStart. The lower bound must be the slice
-		// start index, not 0 — a de below the start makes l.Data[nameStart:de] a
-		// low>high slice expression that panics. Checking de < nameStart guarantees
-		// nameStart <= de <= len(l.Data), so the slice below can never panic.
-		de := nameStart + (int(c) << 1)
-		if de > len(l.Data) || de < nameStart {
-			return nil
-		}
-		b, err := hex.DecodeString(l.Data[nameStart:de])
-		if err != nil {
-			return nil
-		}
-		return &bchain.AddressAliasRecord{Address: address, Name: string(b)}
-	}
-	return nil
-}
+// ENS parsing lives in ens.go.
